providers: split into authenticate() into initiateAuthentication and continueAuthentication
Implemented authenticate() in terms of the latter two.

Added (drumroll please) a userState parameter to initiateAuthentication.
This userState will in future commits be returned by
continueAuthentication in its AuthInfo.

In most cases there's a bit of change in control flow to distinguish the
two cases before the continueAuthentication error-handling, rather than
after.

The actual implementation of the userState persistence is in the next
commit